// src/API/useCreateDepartment.types.ts
export type CreateDeparatmentType = {
  departmentLogo: File | null;
  departmentName: string;
  departmentAbbreviation: string;
  departmentAdminUsername: string;
  departmentAdminEmail: string;
  departmentAdminPassword: string;
  departmentAdminFirstName: string;
  departmentAdminLastName: string;
  departmentAllocatedStorage: number;
};